+++ src/clj/stripe/http.clj
@@ -1,6 +1,7 @@
 (ns stripe.http
   (:require [cheshire.core :as json]
             [clojure.core.async :as a]
+            [environ.core :as e]
             [org.httpkit.client :as http]
             [schema.core :as s]
             [stripe.schema :as ss]
@@ -33,11 +34,17 @@
 
 (s/defn api-token :- (s/maybe ApiToken)
   []
-  *token*)
+  (or *token* (:stripe-secret e/env)))
 
 (defmacro with-token [k & forms]
   `(binding [*token* ~k]
      ~@forms))
+
+(defn use-token!
+  "Permanently sets a base token. The token can still be overridden on
+  a per-thread basis using with-token."
+  [s]
+  (alter-var-root #'*token* (constantly s)))
 
 ;; ## Private