+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.5
 
 - upgrade to schema 0.3.0
+- Add the ability to set the stripe_secret with an environment variable.
 
 ## 0.1.4 (10/08/2014)