+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 0.1.5
+
+- upgrade to schema 0.3.0
+- Add the ability to set the `:stripe-secret` with an environment variable.
+- Change `:stripe-dev-token` to `:stripe-secret`
+
+## 0.1.4 (10/08/2014)
+
+- Fix schema bug in http and token.
+
+## 0.1.3 (10/08/2014)
+
+- Add `http/use-token!`, for permanent token setting
+
 ## 0.1.2 (10/08/2014)
 
 - Relax charge schema until I can get errors going.